Client Service Update

+++ Domain/Entities/WorkRequest.cs
@@ -11,7 +11,7 @@ namespace Domain.Entities
         public string WorkerEmail { get; set; }
         public string ClientEmail { get; set; }
         public string Message { get; set; }
-        public double? Rating { get; set; }
+        public int? Rating { get; set; }
         public bool? IsAccepted { get; set; }
         public bool IsCompleted { get; set; }
     }
